organized builder tests and moved jest files back into test
