+++ test/integration/decorations.integration.test.ts
