+++ test/integration/playsnicely.integration.test.ts
