+++ test/integration/suggestions.integration.test.ts
